move getImprovements to helpers
